(change version) Added new function and options, -l -u which trigger partition label and uuid output.
Either -u or -l will also automatically trigger the short -P partition flag.

To use with full partition output, use: -pul or -pu or -pl

This is one of the last major features I wanted to add, and it's because I'm sick of manually checking for uuid or label in /dev/disk..

Ok, have fun.