1.5.2: new options: -z, triggers output filtering. This is default on for irc client output.
-Z - absolute override, turns off all filtering. This is useful if you are helping someone on irc for example and need to see this 
data.

Fixed a long term bug in the regex that filtered out home name from mount paths in partitions, was greedy, now it just filters out 
only the name and prints the rest of the full path, as it should.

Added twm to desktop support list, that has no --version output so that won't show.