new tarball, new version. This fixes a somewhat obscure issue where for example rpmforge adds apt repos, and also I assume some apt rpm tools might do the same for yum repos. Now inxi will simply note the presence of each type, for apt and yum.
I didnt' do this feature for pacman and pisi, because I've never heard of that, but if necessary, I'll add it.

Now inxi will hopefully correctly note the presence of the various repos and list the files with the repo type correctly
noted for each type.

I also fixed a bug in the yum/rpm type repo detection, inxi failed to properly handle spaces around the = sign, which made
it falsely identify as enabled disabled repos, it tested for only ^enabled=1 whereas also ^enabled = 1 is a possible syntax.

This is now corrected.