New version, tarball. Fixed a small issue that would create a wrong reporting of disk useage if bind mounts are used, ie, multiple binds to a single mount. Now inxi will check a list of the previously used partitions before adding the size of the used space to the total used, if the partition has already been used it will skip it. This was/is a quick and dirty fix, but it's totally fine I believe and should resolve two separate issues:
1. use of bind mount method, where multiple partition names are bound to the same partition
2. accidental dual mounting to the same partition.

partitions section will still show the same data, ie, if bind is used, it will show all
the bind mounts even when they are attached/bound to a partition that is already listed.
This seems useful information, though maybe we can get the key word 'bind' in there somehow,
but for now I won't worry about that issue, that's just a nice to have, not a bug.