New version, new tarball. Another stab at finally fixing the cpu / core count failures on fringe cases. This required fixing some core logic assumptions that are not currently correct, particularly on two cases, some xeon cpus fail to show core id for each core, showing 0 for all of them, second, vm cpus do not show physical ids at all for at least intel, nor do they show core id.
While we can't get HT totally reliable, particularly for vm xeon, since inxi has no way to know in
that case if a core is attached to a physical core or a virtual one, all of them being virtual in that
case, but still inxi is now reporting the correct number of cores, or threads in vm xeons, and is not
showing multicore cpus as single core, which was the main issue.

This required redoing the counter logic for the cpu/core/physical arrays, now they are set independently,
and can handle any of the others not being set, without creating an error or failure condition.

Also added in last check for a certain intel case where core id is 0 but > 1 physical cores exist, that
now also shows the correct cpu / core count.

While this is tested on many data sets of proc cpuinfo, it's still possible there is a fringe case I have
not seen that will trigger yet another unexpected behavior.