New version. Added dynamic wrapping to -G, and also am now wrapping -C per cpu cores speeds, for systems with a lot of them, that will clean up the output.
Added dynamic wrapping to --recommends and -c 94-99.

These are the main things, there's a few smaller issues with -xx output on -N/-n/-i but
those will noly really show with full output and it takes a while to get this stuff stable
so maybe some other time, but it's ok for now.