New version, new tarball, new man page. Finally, after all these years, initial memory/ram support. This feature requires dmidecode, and usually that needs to be run as root.
Significantly improved dmidecode error handling and output, and have as 2.1.90 testing/initial
release basic ram data. 

In subsequent releases, extra info for -x and -xx and -xxx will be added as well to the output.

For those who want to jump on board early for ram data, update your repos, for those who want to
wait for the full featured version, with -x type data, wait for 2.2.0

And that's that.