New Feature, new version, new man page, new tarball. Laptop users should be happy, -B option now shows, if available, battery data. Quite good data for systems with /sys battery data, only rudimentary for systems using dmidecode (BSDs). dmidecode has no current voltage/charge/current supported capacity.
Main row shows charge and condition. Condition shows you have much capacity the
battery currently has vs its design capacity. Charge shows the Wh/percent of
current capacity of battery (NOT the rated design capacity).

-x adds battery vendor/model info, and battery status (like, charging, discharging,
full).

-xx adds battery serial number and voltage information. Note that voltage information
is presented as Current Voltage / Designed minimum voltage.

-xxx adds battery chemistry (like Li-ion), cycles (note: there's a bug somewhere in
that makes the cycle count always be 0, I don't know if that's in the batteries,
the linux kernel, but it's not inxi, just FYI, the data is simply 0 always in all
my datasets so far.

For dmidecode output, the location of the batter is also shown in -xxx