New version, tarball, man page. This closes issue #122. Adds support for including nvme disk capacity in full disk capacity listing. Adds nvme name/serial/firmware revision number. The latter is a new -Dxx output option. Note that as far as I could tell, so far, nvme is the only disk type that has firmware revision data.
Added support for nvme disk temperature as well, that requires the cli tool nvme.

Updated AMD microarchitecture list to be more granular and complete. Added Intel
microarch type. Note that they are releasing a few new microarchitectures soon but I
was not able to find any model numbers for those.