New version, new tarball. This is an attempt at a fix for issue #129
Because I don't want to break existing cpu logic, I just added in a rizen switch,
which will just use cpu_core_count value, then trigger HT output.

This fix may or may not work, but the issue poster vanished and has not followed up.

For now I'm keeping this a Ryzen specific adjustment, but it may be safe to extend
it further, that is, if siblings > 1 && siblings = 2 * cores then it's HT.