Code: big update, added downloader logic, debugged, that works. Added tools for update_me, but not update me itself, which I had to make the downloader tools for. All seems to work.
Created more module files, or libraries, whatever.

Updated inxi_vlaues, programming.

This is a big update, and almost gets us to the point where pinxi
actually does something, in this case, it will be the -U --update
option, which I want to get done quickly for testing, and making it
easy to run pinxi on various machines.

Now that the downloader logic is solid, that will also cover -i (for non
dig users) and -w/-W --weather options, along with the -U update option.