More cleanup, made eval start/end now use conditional to run it, as in: eval $start if $b_log which may save a small amount of cpu/ram time, we'll see.
More optimization tests: conclusion: it makes zero difference
in execution time if you use direct static: ClassName::method() or
$ob_class =ClassName->new()
$ob_class->method();

like, literally none. I also tested pulling the package out of its
wrapper and running it as a set of subs, and it was the same there too,
so apparently perl really does not care re speed which way it's done.

that's good, because I'm going to be using a lot of packages.