First small patch fix, corrected a few issues, one for apt deb822 output formatting, and a small bug for blank files there.
Second, made the --output error message more clear for bad file structure now lists
the 3 requirements: must be full path, must be writeable directory, and must have
a file in it.

Third, another subtle thing, after a lot of research, am trying the MiB GiB format
because it's technically more accurate and less ambiguous than GB, which is used
either to refer to 1000 bite blocks OR to 1024 blocks, depending on the platform etc.

So rather than hope people get it, trying that slightly more wordy format, and maybe
if people wonder what it is.