New version, new man. Bug fixes, feature tweaks.
Bugs fixed:
1. json/xml outputter had a bug in it that made it validate path wrong.
2. -G -xx option: compositor: for gnome-shell had a bug that would make it show
as running when it wasn't, other strings were tripping the match on systems with
gnome-shell installed but not running,
3. Finally fixed bug with manjaro full version distro string, and tweaked output
to show Manjaro Linux instead of given string.

Features added:
1. --no-man - this lets users turn off man installs. Only really useful for -U
from master, since default is off for pinxi  and dev 3 branch.

Man page/help updated to add this option.