New version, new man. Small new enhancements.
1. Added to -s for ipmi, with -x: voltage 12v,5v,3.3v,vbat; for -xx, dimm/soc p1/p2
voltages
2. enhanced wm: feature, needed more filters and protection against redundant data
3. basic apple osx fixes to keep it from crashing, but I'm not spending any more
time on apple junk unless someone pays me for my time, I can't stand the product or
company, it's the total antitheses of freedom or free software, or even openness.
4. openbsd/bsd fixes: openbsd was failing to get cpu flags due to a small oversight
5. -C now shows bits: for the true bits of cpu, not the kernel bits. This is not
a reliable measurement but should be right about 95+ percent of the time, and
basically all of the time for GNU/Linux on Intel/AMD, most of the time for ARM.
When it doesn't know it does not guess, and shows N/A.
6. bsd fix for usb, was running numeric action on string value
7. fixed stderr tool for program_version, now it's hard-coded in program_values
which removes an unneded regex search for every program version test.
8. Mate detection, switched to using mate-sesssion instead of mate-about, the
latter is not getting updated and has the wrong version number on it.