New version, new tarball. New features, bug fixes.
This is a big one.

NEW FEATURES:

1. By Request: Disk vendor is now generally going to be shown. Since this uses
empirical data to grab the vendor name, from the model string, it will not always
find anything. When it fails to find vendor data, no vendor: item will show.

Note that some MMC devices will probably not show vendor data, but that's due to
there being no data that reveals that.

2. Extended -sx volts to also show voltage from lm-sensors if present. Many
systems show no voltage data with lm-sensors, but now if any is found, it
will show, same as impi.

3. Moved to lsblk as primary source for partition/unmounted filesystem, uuid, and
label data.

Falls back to previous methods if lsblk does not return data. Some lsblk do not
show complete data unless super user as well.

4. Refactored code to be more logical and clear.

5. Added for OpenBSD -r: /etc/installurl file.

BUG FIXES:

1. CRITICAL: /sys/block/xxx/device/model is in some cases truncating the disk
model name to 16 characters. This is not  an inxi bug, it's a bug with /sys itself.

To fix this, inxi now uses for GNU/Linux /dev/disk/by-id data which does not
ever do this truncation. It's also faster I believe to read that directory
once, filter the results, then use the data for vendor/model/serial.

this was also part of the disk vendor data feature.

2. Openbsd networking fix. Was not showing IF data, now it does.

3. Fixed bug with unmounted where sometimes md0 type partitions would show
even though they are in a raid array.

4. Fixed disk rev, now it searches for 3 different files in /sys to get that data.

5. Fixed bug with very old systems, with sudo 1.6 or older, for some reason that
error did not get redirected to /dev/null, so now only using sudo -n after explicit
version test, only if 1.7 or newer.

6. Fixed a few null results in fringe cases for graphics. Resolution now shows
NA for Hz if no hz data found. This was only present on a fringe user case
which is unlikely to ever impact normal X installations.

7. Fixed BSD L2 cache, was showing MiB instead of KiB, wrong math.