New version, new man. Bug fixes, feature updates.
The main reason to release this earlier than I had hoped was because of the /sys
permission change for serial/uuid file data. The earlier we can get this fix out,
the better for end users, otherwise they will think they have no serial data when
they really do.

FIXES:
1. this bug just came to my attention, apparently the (I assume) kernel people
decided for us that we don't need to see our serial numbers in /sys unless we are
root. This is an unfortunate but sadly predictable event. To work around this
recent change (somewhere between 4.14 and 4.15 as far as I can tell), inxi -M and
-B now check for root read-only and show <root required> if the file exists but
is not user readable. I wish, I really wish, that people could stop changing stuff
for no good reason, but that's out of my control, all I can do is adjust inxi to
this reality. But shame on whoever decided that was a good idea.

This is not technically an inxi bug, but rather a regression, since it's caused
by a change in /sys permissions, but users would see it as a bug so I consider
this an important fix.

Note that the new /sys/class/dmi/id permissions result in various possible things:
1. serial/uuid file is empty but exists and is not readable by user
2. serial/uuid file is not empty and exists and is not readable by user
3. serial/uuid file does not exist
4. serial/uuid file exists, is not empty, and is readable by root

Does this change make your life better? It doesn't make mine better, it makes
it worse. Consider filing a bug report against whoever allowed this regression
is my suggestion.

BUGS:
1. A weather bug could result in odd or wrong data showing in weather output, this
was due to a mistake in how the weather data was assembled internally. This error
could lead to large datastore files, and odd output that is not all correct.

2. More of an enhancement, but due to the way 'v' is used in version numbers,
the program_version tool in some cases could have sliced out a 'v' in the wrong
place in the version string, and also could have sliced out legitimate v values.

This v issue also appeared in bios version, so now the new rule for program_version
and certain other version results is to trim off starting v if and only if it is
followed by a number.

FEATURES:
1. Added in OpenBSD support for showing machine data without having to use dmidecode.
This is a combination of systcl -a and dmesg.boot data, not very good quality data
sources, but it is available as user, and it does work. Note that BIOS systems
are the only ones tested, I don't know what the syntax for UEFI is for the field
names and strings. Coming soon is Battery and Sensors data, from the same sources.

Sadly as far as I know, OpenBSD is the only BSD that has such nice, usable (well,
ok, dmesg.boot data is low quality strings, not really machine safe) data. I
have no new datasets from the other BSDs so I don't know if they have decided to
copy/emulate this method.

2. By request, and this was listed in issue #134, item no. 1, added in weather
switchable metric/imperial output. Also added an option, --weather-unit and
configuration item: WEATHER_UNIT with possible values: cf|fc|c|f. The 2nd of
two in cf/fc goes in () in the output. Note that windspeed is m/s or km/h as metric,
inxi shows m/s as default for metric and (km/h as secondary). Also fixed -w
observation date to use local time formatting. That does not work in -W so it shows
the default value.

3. Updated man to show new WEATHER_UNIT config option, and new --weather-unit
option. Also fixed some other small man glitches that I had missed.