New version, new tarball.
This version is very peaceful, no big changes, just a few fixes and small new
features added.

This version corrects a few small glitches reported by users, and adds basic support
for disk speed report. Note that this is not as accurate as I'd like, it tries, but
there is not a lot of data to be had. Limits of disk speed seems to be, roughly:
1. most speed is reported as max board can do, not max drive can support
2. usually when speed is reported as lower than max board speed, it's correct, but,
as usual, exceptions to this were found during testing.
3. usually if drive is faster than board speed, it reports board speed, but, again,
exceptions to this rule were found during testing.

However, with this said, it's usually more or less right, at least right in terms
of the fastest speed you can expect to get with your board. NVMe was also supported,
that's much more complicated because NVMe has >= 1 lane, and each lane has up and
down data. The reported speed is max in one direction, and is a function of the
PCIe 1,2 20% overhead, and PCIe 3,4,5 ~1.5% overhead. inxi shows the actual usable
data rate, not the GT/s rate, which is the total transfers per second the unit
supports.

So due to the unreliable nature of the data, this is only a -xx option. There is
also in general no data for USB, and none for mmcblk (sd cards usually).

This feature may be enhanced with a C Perl XS library in the future, we'll see how
that goes.

FIXES:
1. corrected an issue where a networking card of type Bridge failed to be detected.
This is now handled. This was a PCI type I'd never seen before, but it exists, and
a user had it, so now it will work as expected for this type.
2. changed the default units in weather to be m (metric) imperial (i). While this is
not very intuitive for me, it's easier to explain I think. The previous c / f
syntax is supported internally, and inxi will just translate c to m and f to i, so
it doesn't matter which is or was used on a config file or with the --weather-unit
option.
3. BSD uptime had a parsing glitch, there was a spelling variant I'd never seen in
GNU/Linux that broke the regex. This is corrected now.
4. Fixed a few small man page glitches, some ordering stuff, nothing major.
5. Fixed BSD hostname issues. There was a case where a setup could have no hostname,
inxi did not handle that correctly. This fix would have applied to gnu/linux as
well.
6. Fixed a few bsd, openbsd mostly, dm detections, there is a secondary path in
OpenBSD that was not checked. This also went along with refactoring the dm logic
to be much more efficient and optimized.
7. Fine tuned dmidecode error message.
8. Fixed PCI ID issue, it was failing to catch a certain bridged network type.
9. A more global fix for unhandled tmpfs types, in this case, shm, but added a
global test that will handle all tmpfs from now on, and exclude that data from
-p reports.

NEW FEATURES:
1. First attempt to add basic disk speed (Gb/s). Supported types: ATA, NVMe. No
speed data so far handled or found: mmcblk; USB. Also possibly older /dev/hda
type devices (IDE bus) may not get handled in all cases. This may get more work
in the future, but that's a long ways off. This case oddly was one where BSDs had
support for basic disk speed reports before GNU/Linux, but that was really just
because it was part of a single data line that inxi parsed for disk data anyway
with BSDs.
2. Man items added for -Dxx disk speed options.