New version, man page.  Features, bugs, fixes!
Bugs:
1. Color selector accepted '' as a value, which it would then write to config
file, creating errors since it's not an integer value.
2. Corrected distro id error for last fallback case, making the distro ID out
of the filename itself, that was missing the assignment to $distro.
3. mmcblk0 was showing up as an unmounted partition, due to failing to filter
mmcblk[0-9] in unmounted.

Fixes:
1. Added missing compositor kwin_wayland to compositor detections
2. For -M, on laptops, sometimes Type: would duplicate in Chassis: type: which
looks silly, so now it checks to make sure the two values are different before
using the Chassis: type: data.
3. -D disk vendor, added GALAX, fixed Toshiba, which sometimes occurs other than
start of disk id string, so now it checks the whole string. This seems particularly
common in nvme devices from Toshiba. This is the only vendor I have found that
puts the vendor string later in the device id string.
4. Added protection against unreadable but present /etc/issue. This was caused
by a now fixed bug in OpenSuSe, which symbolically linked to create /etc/issue
from /var/run/issue, but with 600 permissions, root read only, that is. Note that
this bug has since been fixed (now has the correct 644 permissions), but I figured
better safe than sorry in case anyone else decides that's a good idea in the future.
Now only sends to reader if readable.
5. Related to 4, made reader not exit on failure, now just prints error message and
keeps going.
6. Upped maximum distro string length to 60, from 50. AntiX for example was coming
in at 48, so I decided to add some safe room now that inxi does dynamic sizing, it
is not a big problem having very long distro id strings.

Enhancements:
1. Added basic /proc data parser to debugger. Can't get all the data or files because
it's simply too big, but grabs the basics.
2. Added vcgencmd for some ARM rasberry pi debugging.
3. ARM: add model if not found in /proc/cpuinfo, or if different.
4. Added Tdie cpu sensor type, this is coming soon in latest kernels, so catching
it early. Tdie will replace k10-temp sensor item temp1.
5. Added --admin extra data option, and first set of extra data, -C, which will
show CPU Errata (bugs), family, model-id, stepping (as hex (decimal) or hex if less
than or equal to 9), microcode (as hex).
6. Battery: added with -x option, if found, attached battery driven devices, like
wifi keyboard, mouse. If upower is present, will also try to show battery charge
percent for those devices. Note that -B only shows the Device-X items if -x is used,
and will not show anything in -F unless there is a system, not device, battery
present, or if -Fx is used and there is a Device battery detected.
Added upower to recommends.
7. Basic -Dxxx disk rotation speeds added. Requires udevadm. Not all spinning disks
show rotation speeds, and it depends on udevadm, so if no rotation found, it shows
nothing.
8. Added explicit Arco Linux and Antergos distro ID support. This requires more
checks, but in theory, both should now show Arco Linux or Antergos instead of default
'Arch Linux' as before, plus extra data if found, like version.