New version, man page. New features and fixes!
Bugs:
1. -I line, sometimes running in showed sudo. This is hopefully now corrected.

Fixes:
1. CPU architectures, small reordering based on hopefully more reliable data
source, but these are hard to find conclusively.
2. -S Distro id: switched ordering of prefered os-release sources, PRETTY_NAME
is not being used consistently, too many distros leave out the distro id found
in VERSION, so now it uses NAME + VERSION if both are there, then PRETTY_NAME
as a fallback. That reverses how it was, but it will provide better results for
most distros. Distros that did this properly to begin with should see no change.
3. Now that inxi is basically debugged and working, I've removed the output of
'inxi' from the -t lines. It remains for the pinxi branch however so you can
see how many resources pinxi uses to run.
4. ipmi sensors data are proving to be as random as lm-sensors. Added another
alternate syntax for sensors.
5. CPU: found an alternate syntax, again, for IPMI and sensors data, added
support, I hope, for that.

Enhancements:
1. Added /proc debugger tool to debugger. Due to oddities with how the /proc file
system is created, it will only run as user, not root, unless the --proc flag is
used. More programs added to debugger commands.
2. More disk vendor strings added, fine tuning of vendor detections. There is a
tendency in NVMe disk names to put the vendor name in the middle of the string.
That is now handled for a few key vendors.
3. Added basic ARM SOC and server support. This will require more work in the future
because the syntax used varies significantly device to device, but the featuers
are now in place to add that support. Most SBC ARM devices should now at least show
the model and details data in machine data, and some will show -G -A -N data as
well.
4. ARM CPU: added first attempt to show the cpu variant as well as the more
generic ARM data. This shows 1 or more variants, some ARM devices have two different
cpu cores running at different speeds. Odroid for example.
5. Added system 'base:' data for -Sx, that modifies Distro: in supported cases.
Currently only Mint and MX/AntiX supported because each specific distro must be
handled explicitly using empirical file based data tests. I decided against showing
this for rolling release, since really everyone knows that Antergos is made from
Arch Linux, so showing that does not provide much useful information, whereas
showing the Ubuntu version Mint was made from does.

Note that several derived distros are changing how they use os-release, so the
tools had to be revised to be more dynamic, which is a pain, and makes it even
more empirical and less predictable to print what should be trivially easy to
gather distro and derived source data.

If your distro is not in this list and you want the base data to be present, please
supply a --debug 22 dataset so I can check all the files required to make the
detection work. If your distro has changed methods, please note which methods
were used in the past, and which are used now.
6. Added Armbian distro detection, that's tricky. Added Rasbpian detections.
Added improved Antergos, Arco, and maybe Chakra, Arch detections.
7. Big one: Hardware RAID basic support added. Note that each vendor, and
unfortunatley, often each product line, has its own raid status and drive
reporting tools, which makes adding the actual drive/raid/status report part
very time consuming to add. I may only support this if a certain software maker's
raid tools are installed because they are much simpler to parse, but for now,
it only shows the presence of the raid device itself, not disks, raid status, etc.